mysite/requirements.txt
Django==2.1.4
gunicorn==19.9.0
whitenoise==4.1.2
Babel==2.6.0
dj-database-url==0.5.0
django-phonenumber-field==2.1.0
phonenumbers==8.10.2
pytz==2018.7
psycopg2==2.7.2